fix(eval): cast actions to float32 before numpy conversion
Policies that output bfloat16 tensors (e.g. X-VLA) crash during
evaluation because numpy does not support bfloat16. Explicitly
cast to float32 when converting actions to numpy.

Fixes #2956

+++ src/lerobot/scripts/lerobot_eval.py
@@ -181,8 +181,8 @@ def rollout(
         action_transition = env_postprocessor(action_transition)
         action = action_transition[ACTION]
 
-        # Convert to CPU / numpy.
-        action_numpy: np.ndarray = action.to("cpu").numpy()
+        # Convert to CPU / numpy. Cast to float32 first since numpy doesn't support bfloat16.
+        action_numpy: np.ndarray = action.to("cpu", dtype=torch.float32).numpy()
         assert action_numpy.ndim == 2, "Action dimensions should be (batch, action_dim)"
 
         # Apply the next action.
